gyatword-backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
supabase==2.11.0
fastapi-auth0==0.5.0
python-dotenv==1.0.1
requests==2.32.3